requirements.txt
accelerate==0.18.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
cohere==4.19.2
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
datasets==2.11.0
dill==0.3.6
exceptiongroup==1.1.2
faiss-cpu==1.7.4
Faker==19.2.0
fastavro==1.8.2
fastcore==1.5.29
favicon==0.7.0
filelock==3.12.2
filetype==1.2.0
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
htbuilder==0.6.1
huggingface-hub==0.13.4
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.266
langsmith==0.0.22
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==10.0.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==9.1.0
plotly==5.15.0
protobuf==4.23.4
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.0
pymdown-extensions==10.1
Pympler==1.0.1
PyMuPDF==1.22.5
pyparsing==3.0.9
pypdf==3.15.1
PySocks==1.7.1
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
responses==0.18.0
rich==13.5.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.11.2
sentence-transformers
sentencepiece==0.1.99
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
st-annotated-text==4.0.0
streamlit==1.25.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-chat==0.1.1
streamlit-embedcode==0.1.2
streamlit-extras==0.3.0
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.28.1
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.9.2
urllib3==2.0.4
validators==0.20.0
wsproto==1.2.0
xxhash==3.3.0
yarl==1.9.2